Fixed select.php and made insert.php

+++ d09/ex04/insert.php
@@ -1,5 +1,8 @@
 
 <?php
-	file_put_contents('list.csv', $_GET['id'] . ';' . $_GET['value'] .
-		PHP_EOL, FILE_APPEND);
+	if (file_exists("list.csv") && isset($_GET['id']) && isset($_GET['value']))
+	{
+		file_put_contents('list.csv', $_GET['id'] . ';' . $_GET['value'] .
+			PHP_EOL, FILE_APPEND);
+	}
 ?>
